fix(master): 16/05/2024 08:54

+++ src/main/java/pokedex/repository/PokemonRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Component;
 import pokedex.entity.Pokemon;
 
 @Component
-public interface PokemonRepository extends CrudRepository<Pokemon,Long> {
+public interface PokemonRepository extends CrudRepository<Pokemon,Integer> {
 }
