+++ src/main/java/pokedex/repository/TrainerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Component;
 import pokedex.entity.Trainer;
 
 @Component
-public interface TrainerRepository extends CrudRepository<Trainer,Long> {
+public interface TrainerRepository extends CrudRepository<Trainer,Integer> {
 }
